requirements.txt
Flask==2.3.3
Flask-Cors==5.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
websockets==11.0.3
mysql-connector-python==9.0.0
